requirements.txt
flask>=3.1.0
python-dotenv>=1.0.0
requests>=2.32.0
nltk>=3.9.0
transformers>=4.50.0
numpy>=2.0.0
pandas>=2.0.0
scikit-learn>=1.6.0
torch>=2.0.0
openai>=1.0.0
anthropic>=0.10.0